refactor: refactor

+++ JSSoft.Communication/InstanceContext.cs
@@ -20,18 +20,18 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using JSSoft.Library.ObjectModel;
-using JSSoft.Library.Threading;
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace JSSoft.Communication;
 
 sealed class InstanceContext : IInstanceContext, IPeer
 {
-    private readonly Dictionary<IPeer, PeerDescriptor> _descriptorByPeer = new();
+    private readonly ConcurrentDictionary<IPeer, PeerDescriptor> _descriptorByPeer = new();
     private readonly PeerDescriptor _descriptor = new();
     private readonly ServiceContextBase _serviceContext;
 
@@ -41,19 +41,19 @@ sealed class InstanceContext : IInstanceContext, IPeer
         ID = Guid.NewGuid();
     }
 
-    public async Task InitializeInstanceAsync()
+    public async Task InitializeInstanceAsync(CancellationToken cancellationToken)
     {
         var query = from item in _serviceContext.ServiceHosts
                     where ServiceContextBase.IsPerPeer(_serviceContext, item) != true
                     select item;
         foreach (var item in query)
         {
-            var (service, callback) = await Task.Run(() => _serviceContext.CreateInstance(item, this));
+            var (service, callback) = await Task.Run(() => _serviceContext.CreateInstance(item, this), cancellationToken);
             _descriptor.AddInstance(item, service, callback);
         }
     }
 
-    public async Task ReleaseInstanceAsync()
+    public async Task ReleaseInstanceAsync(CancellationToken cancellationToken)
     {
         var isServer = ServiceContextBase.IsServer(_serviceContext);
         var query = from item in _serviceContext.ServiceHosts.Reverse()
@@ -61,12 +61,12 @@ sealed class InstanceContext : IInstanceContext, IPeer
                     select item;
         foreach (var item in query)
         {
-            var (service, callback) = await Task.Run(() => _descriptor.RemoveInstance(item));
+            var (service, callback) = await Task.Run(() => _descriptor.RemoveInstance(item), cancellationToken);
             _serviceContext.DestroyInstance(item, this, service, callback);
         }
     }
 
-    public async Task<PeerDescriptor> CreateInstanceAsync(IPeer peer)
+    public PeerDescriptor CreateInstance(IPeer peer)
     {
         var peerDescriptor = new PeerDescriptor();
         foreach (var item in _serviceContext.ServiceHosts)
@@ -74,7 +74,7 @@ sealed class InstanceContext : IInstanceContext, IPeer
             var isPerPeer = ServiceContextBase.IsPerPeer(_serviceContext, item);
             if (isPerPeer == true)
             {
-                var (service, callback) = await Task.Run(() => _serviceContext.CreateInstance(item, peer));
+                var (service, callback) = _serviceContext.CreateInstance(item, peer);
                 peerDescriptor.AddInstance(item, service, callback);
             }
             else
@@ -83,19 +83,20 @@ sealed class InstanceContext : IInstanceContext, IPeer
                 peerDescriptor.AddInstance(item, service, callback);
             }
         }
-        _descriptorByPeer.Add(peer, peerDescriptor);
+        _descriptorByPeer.TryAdd(peer, peerDescriptor);
         return peerDescriptor;
     }
 
-    public async Task DestroyInstanceAsync(IPeer peer)
+    public void DestroyInstance(IPeer peer)
     {
-        var peerDescriptor = _descriptorByPeer[peer];
+        if (_descriptorByPeer.TryRemove(peer, out var peerDescriptor) == false)
+            return;
         foreach (var item in _serviceContext.ServiceHosts.Reverse())
         {
             var isPerPeer = ServiceContextBase.IsPerPeer(_serviceContext, item);
             if (isPerPeer == true)
             {
-                var (service, callback) = await Task.Run(() => peerDescriptor.RemoveInstance(item));
+                var (service, callback) = peerDescriptor.RemoveInstance(item);
                 _serviceContext.DestroyInstance(item, peer, service, callback);
             }
             else
@@ -104,7 +105,7 @@ sealed class InstanceContext : IInstanceContext, IPeer
             }
         }
         peerDescriptor.Dispose();
-        _descriptorByPeer.Remove(peer);
+
     }
 
     public object? GetService(Type serviceType)
@@ -115,8 +116,6 @@ sealed class InstanceContext : IInstanceContext, IPeer
                     select service;
         return query.SingleOrDefault();
     }
-
-    public Dispatcher Dispatcher => _serviceContext.Dispatcher;
 
     public Guid ID { get; }
 }
